requirements.txt
flask
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain
pypdf
google-generativeai
-e .
ipykernel
